+++ test/fixtures/fake-provider/index.js
@@ -1,5 +1,4 @@
 exports.name = 'test-provider'
-exports.pattern = '/:id'
 exports.type = 'provider'
 exports.status = { version: '0.0.0' }
 exports.Controller = require('./controller')