+++ test/helpers/Cache.js
@@ -1,6 +1,7 @@
 const Cache = require('../../src/cache')
 const local = require('../../src/local')
-var config = { logfile: __dirname + '/../test.log' }
+const path = require('path')
+var config = { logfile: path.join(__dirname, '/../test.log') }
 
 // init the koop log based on config params
 const Logger = require('koop-logger')