+++ test/index.js
@@ -1,14 +1,16 @@
 /* global describe, it */
 
-var provider = require('../fixtures/fake-provider')
-const Koop = require('../../src/')
+var provider = require('./fixtures/fake-provider')
+const Koop = require('../src/')
 const koop = new Koop()
+const should = require('should') // eslint-disable-line
 
 describe('Index tests for registering providers', function () {
   describe('can register a provider', function () {
     it('should register successfully', function () {
       koop.register(provider)
-      koop.server._router.stack.length.should.equal(17)
+      const routeCount = (koop.server._router.stack.length)
+      routeCount.should.equal(17)
     })
   })
 })